Puno dosta jako promjena

+++ android/app/src/main/java/hr/fer/projekt/smartAgriculture/model/MeasurementModel.kt
@@ -6,10 +6,10 @@ data class MeasurementModel(
     val id: Long,
     val device: DeviceModel,
     val time: Date,
-    val airHumidity: Float,
-    val soilHumidity: Float,
-    val airTemperature: Float,
-    val soilTemperature: Float,
+    val airHumidity: Double,
+    val soilHumidity: Double,
+    val airTemperature: Double,
+    val soilTemperature: Double,
     val pressure: Double,
     val luminosity: Double
 )
